requirements.txt
yt-dlp>=2024.0.0
ffmpeg-python>=0.2.0
